# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Volleyball Tryout Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players
      parameters:
        - name: position
          in: query
          schema:
            type: string
        - name: missingMeasurements
          in: query
          schema:
            type: boolean
        - name: checkedIn
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Create a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/import-csv:
    post:
      operationId: importPlayersCsv
      tags: [players]
      summary: Import players from CSV data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CsvImport"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
...
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get a player by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerDetail"
        "404":
          description: Not found
    patch:
      operationId: updatePlayer
      tags: [players]
      summary: Update a player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdate"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Not found
    delete:
      operationId: deletePlayer
      tags: [players]
      summary: Delete a player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # Evaluations
  /evaluations:
    get:
      operationId: listEvaluations
      tags: [evaluations]
      summary: List evaluations
      parameters:
        - name: playerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of evaluations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Evaluation"
    post:
      operationId: upsertEvaluation
      tags: [evaluations]
      summary: Create or update an evaluation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluationInput"
      responses:
        "200":
          description: Upserted evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evaluation"
  /evaluations/{id}:
    patch:
      operationId: updateEvaluation
      tags: [evaluations]
      summary: Update an evaluation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluationUpdate"
      responses:
        "200":
          description: Updated evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evaluation"
        "404":
          description: Not found

  # Rankings
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List coach notes
      parameters:
        - name: playerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoachNote"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a coach note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachNote"

  # Coaches
...
  /notes/{id}:
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a coach note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachNote"
        "404":
          description: Not found
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a coach note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # Sync
...
components:
  schemas:
...
    Player:
      type: object
      required: [id, jerseyNumber, name, position, checkedIn, createdAt, updatedAt]
      properties:
        id:
          type: integer
        jerseyNumber:
          type: string
        name:
          type: string
        position:
          type: string
          enum: [Setter, OutsideHitter, MiddleBlocker, Opposite, Libero]
        checkedIn:
          type: boolean
        heightInches:
          type: ["number", "null"]
        standingReachInches:
          type: ["number", "null"]
        verticalJumpInches:
          type: ["number", "null"]
        overallScore:
          type: ["number", "null"]
        positionScore:
          type: ["number", "null"]
        potentialScore:
          type: ["number", "null"]
        physicalScore:
          type: ["number", "null"]
          description: "Percentile-based score (0–10) from height, reach, and vertical"
        confidenceScore:
          type: ["number", "null"]
          description: "Data completeness score (0–10): how much eval data is available"
        flags:
          type: ["array", "null"]
          items:
            type: string
          description: "Auto-generated player flags (e.g. High Potential, Raw Athlete)"
        rankOverall:
          type: ["integer", "null"]
        rankPosition:
          type: ["integer", "null"]
        rankLocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PlayerDetail:
      allOf:
        - $ref: "#/components/schemas/Player"
        - type: object
          properties:
            evaluations:
              type: array
              items:
                $ref: "#/components/schemas/Evaluation"
            notes:
              type: array
              items:
                $ref: "#/components/schemas/CoachNote"
    PlayerInput:
      type: object
      required: [jerseyNumber, name, position]
      properties:
        jerseyNumber:
          type: string
        name:
          type: string
        position:
          type: string
          enum: [Setter, OutsideHitter, MiddleBlocker, Opposite, Libero]
        checkedIn:
          type: boolean
        heightInches:
          type: ["number", "null"]
        standingReachInches:
          type: ["number", "null"]
        verticalJumpInches:
          type: ["number", "null"]
    PlayerUpdate:
      type: object
      properties:
        jerseyNumber:
          type: string
        name:
          type: string
        position:
          type: string
          enum: [Setter, OutsideHitter, MiddleBlocker, Opposite, Libero]
        checkedIn:
          type: boolean
        heightInches:
          type: ["number", "null"]
        standingReachInches:
          type: ["number", "null"]
        verticalJumpInches:
          type: ["number", "null"]
        rankLocked:
          type: boolean
    CsvImport:
      type: object
      required: [csvData]
      properties:
        csvData:
          type: string
    ImportResult:
      type: object
      required: [imported, updated, errors]
      properties:
        imported:
          type: integer
        updated:
          type: integer
        errors:
          type: array
          items:
            type: string
...
    Evaluation:
      type: object
      required: [id, playerId, category, skill, score, createdAt, updatedAt]
      properties:
        id:
          type: integer
        playerId:
          type: integer
        category:
          type: string
          enum: [universal, position]
        skill:
          type: string
        score:
          type: number
          minimum: 1
          maximum: 10
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EvaluationInput:
      type: object
      required: [playerId, category, skill, score]
      properties:
        playerId:
          type: integer
        category:
          type: string
          enum: [universal, position]
        skill:
          type: string
        score:
          type: number
          minimum: 1
          maximum: 10
        notes:
          type: string
    EvaluationUpdate:
      type: object
      properties:
        score:
          type: number
          minimum: 1
          maximum: 10
        notes:
          type: string
...
    CoachNote:
      type: object
      required: [id, playerId, content, createdAt, updatedAt]
      properties:
        id:
          type: integer
        playerId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    NoteInput:
      type: object
      required: [playerId, content]
      properties:
        playerId:
          type: integer
        content:
          type: string
    NoteUpdate:
      type: object
      properties:
        content:
          type: string
...