# lib/api-zod/src/generated/types/playerDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Volleyball Tryout Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CoachNote } from './coachNote';
import type { Evaluation } from './evaluation';
import type { Player } from './player';

export type PlayerDetail = Player & {
  evaluations?: Evaluation[];
  notes?: CoachNote[];
};